Refactor results into two modules to improve import situation
