+++ src/prefect/client/schemas/actions.py
@@ -46,7 +46,7 @@ from prefect.utilities.collections import listrepr
 from prefect.utilities.pydantic import get_class_fields_only
 
 if TYPE_CHECKING:
-    from prefect.results import ResultRecordMetadata
+    from prefect.result_records import ResultRecordMetadata
 
 R = TypeVar("R")
